+++ discovery-service/src/main/java/com/example/discoveryservice/DiscoveryServiceApplication.java
@@ -1,4 +1,4 @@
-package com.example.ecommerce;
+package com.example.discoveryservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -10,10 +10,10 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
  */
 @EnableEurekaServer
 @SpringBootApplication
-public class EcommerceApplication {
+public class DiscoveryServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(EcommerceApplication.class, args);
+		SpringApplication.run(DiscoveryServiceApplication.class, args);
 	}
 
 }
