+++ discovery-service/src/test/java/com/example/discoveryservice/EcommerceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.ecommerce;
+package com.example.discoveryservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
